requirements.txt
aws-cdk-lib==2.144.0
constructs==10.1.162
PyYAML==6.0
cdk-nag==2.27.178